WIP rails components support
It would be nicer if this lived in `app/helpers` and got autoloaded, but
I cannot get that to work, so don't care for the alpha.

To include in a consuming app, add `helper GovukFrontendAlpha::ApplicationHelper`
 to ApplicationController, and invoke in a view like:

 ```
 <%= govuk_component("button", {
  text: 'My Button',
}) %>
```

This was tested against a manually ERB-converted component template, next
step convert them using meta-template, but this proves the helper approach and
that path loading works.