Drop macro definition from components
Macros are a pretty nunjucks/jinja specific structure, which don't have
equivilents in all the languages we want to support. It also doesn't work
well with Fractal, which expects non macro component templates.

Treat the macro syntax as something package-specific for nunjucks output,
and drop it from the source definition and move the arguments to the
component config, for use in packaging and potentially documentation.

The config structure for arguments is just the names of variables, so
won't support default arguments, or documentation describing the arg.
These are things we can add in the future and update the format, keep
it simple for now.

Looking up config is based on naming convention, template file name to
config file. This is a little messy as we have config per variation,
rather than one canonical config per component, which contains variations,
or using collections. Thats being updated by Gemma on another branch, so
i'm ignoring that for now and updating the configs that make the components
exactly, which does the trick.

The args from config are used in the npm packaging step to rebuild the
macro structure. So the nunucks component output is the same (ignoring
whitespace).

The gulp implementation is pretty sloppy, and this isn't a proper build
step. It should be something more like the transpiler approach, applied
per format, not as part of a particular packing step, as all packages
currently get all templates. This commit is a proof of concept initially.