Move nunjuck macro wrap to component build step
Move it out of the npm packaging step and make it part of the template
building, so it can apply to any template format.

For now only apply it to nunjucks, but move that code into the nunjucks
transpiler.

There's quite a bit of duplication between to the transplier runners in
the template and component build step now. Its worth refactoring once
the component and template transpiling work has settled down a bit.