Quick go at a rails component helper
Only deals with converting the default arugments (rails will
error if referenced when not set, and most arguments so far
are optional). Doesn't touch the main conversion work, which
we're hoping to do with meta-template.

Also wary of the manual erb construction going on, might be
better to add that as part of the nunjucks input before
the meta-template conversion, so the construction is only
in one language, not potetntially more